+++ app/src/main/java/com/growthook/aos/data/datasource/remote/SeedDataSource.kt
@@ -3,6 +3,7 @@ package com.growthook.aos.data.datasource.remote
 import com.growthook.aos.data.model.request.RequestSeedMoveDto
 import com.growthook.aos.data.model.request.RequestSeedPostDto
 import com.growthook.aos.data.model.response.ResponseDto
+import com.growthook.aos.data.model.response.ResponseGetCaveSeedsDto
 import com.growthook.aos.data.model.response.ResponseMoveSeedDto
 import com.growthook.aos.data.model.response.ResponsePostSeedDto
 
@@ -12,4 +13,6 @@ interface SeedDataSource {
     suspend fun moveSeed(seedId: Int, request: RequestSeedMoveDto): ResponseMoveSeedDto
 
     suspend fun postSeed(caveId: Int, request: RequestSeedPostDto): ResponsePostSeedDto
+
+    suspend fun getCaveSeeds(caveId: Int): ResponseGetCaveSeedsDto
 }
