+++ app/src/main/java/com/growthook/aos/data/service/CaveService.kt
@@ -3,7 +3,6 @@ package com.growthook.aos.data.service
 import com.growthook.aos.data.model.request.RequestCaveModifyDto
 import com.growthook.aos.data.model.request.RequestCavePostDto
 import com.growthook.aos.data.model.response.ResponseDto
-import com.growthook.aos.data.model.response.ResponseGetCaveSeedsDto
 import com.growthook.aos.data.model.response.ResponseGetCavesDto
 import com.growthook.aos.data.model.response.ResponseGetDetailCaveDto
 import com.growthook.aos.data.model.response.ResponsePostCaveDto
@@ -43,9 +42,4 @@ interface CaveService {
         @Path("memberId") memberId: Int,
         @Body request: RequestCavePostDto,
     ): ResponsePostCaveDto
-
-    @GET("api/v1/cave/{caveId}/seed/list")
-    suspend fun getCaveSeeds(
-        @Path("caveId") caveId: Int,
-    ): ResponseGetCaveSeedsDto
 }
