+++ app/src/main/java/com/growthook/aos/data/service/SeedService.kt
@@ -3,10 +3,12 @@ package com.growthook.aos.data.service
 import com.growthook.aos.data.model.request.RequestSeedMoveDto
 import com.growthook.aos.data.model.request.RequestSeedPostDto
 import com.growthook.aos.data.model.response.ResponseDto
+import com.growthook.aos.data.model.response.ResponseGetCaveSeedsDto
 import com.growthook.aos.data.model.response.ResponseMoveSeedDto
 import com.growthook.aos.data.model.response.ResponsePostSeedDto
 import retrofit2.http.Body
 import retrofit2.http.DELETE
+import retrofit2.http.GET
 import retrofit2.http.POST
 import retrofit2.http.Path
 
@@ -28,4 +30,9 @@ interface SeedService {
         @Path("caveId") caveId: Int,
         @Body request: RequestSeedPostDto
     ): ResponsePostSeedDto
+
+    @GET("api/v1/cave/{caveId}/seed/list")
+    suspend fun getCaveSeeds(
+        @Path("caveId") caveId: Int,
+    ): ResponseGetCaveSeedsDto
 }
