+++ app/src/main/java/com/growthook/aos/domain/repository/SeedRepository.kt
@@ -1,5 +1,7 @@
 package com.growthook.aos.domain.repository
 
+import com.growthook.aos.domain.entity.Insight
+
 interface SeedRepository {
 
     suspend fun deleteSeed(seedId: Int): Result<Unit>
@@ -7,4 +9,6 @@ interface SeedRepository {
     suspend fun moveSeed(seedId: Int, toMoveCaveId: Int): Result<Unit>
 
     suspend fun postSeed(caveId: Int, insight: String, memo: String, source: String, url: String, goalMonth: Int): Result<Unit>
+
+    suspend fun getCaveSeeds(caveId: Int): Result<List<Insight>>
 }
